fix column type

+++ database/migrations/2020_12_31_083225_create_majors_table.php
@@ -16,7 +16,7 @@ class CreateMajorsTable extends Migration
         Schema::create('majors', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('last_students');
+            $table->tinyInteger('last_students');
             $table->timestamps();
         });
     }
